It's actually coprimes
